requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.21
numpy==1.24.3
pygame==2.5.2
Pillow==10.2.0
tk==0.1.0
